requirements.txt
# Core Framework
streamlit>=1.30.0

# LLM Integration
google-genai>=0.2.0

# Database
# sqlite3 is built-in to Python

# Authentication
bcrypt>=4.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Data Validation
pydantic>=2.0.0

# ML & Analytics
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.44.0

# Utilities
python-dotenv>=1.0.0

# Optional: Embeddings (if using semantic matching)
# sentence-transformers>=2.2.0
# faiss-cpu>=1.7.4

# Optional: Visualization
plotly>=5.17.0

# Development Tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
